+++ neoforge/src/main/resources/crash_assistant-NeoForge.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "dev.kostromdan.mods.crash_assistant.forge.mixin",
+  "package": "dev.kostromdan.mods.crash_assistant.NeoForge.mixin",
   "compatibilityLevel": "JAVA_21",
   "minVersion": "0.8",
   "client": [
